+++ console-ui/src/flow-bpmn/render/index.ts
@@ -1,6 +1,0 @@
-import RuleForgeRenderer from './RuleForgeRenderer';
-
-export default {
-    __init__: ['ruleForgeRenderer'],
-    ruleForgeRenderer: ['type', RuleForgeRenderer]
-};