+++ server/lib/ruleforge-datasource/src/main/java/com/ruleforge/datasource/mapper/DatasourceEntityMappingMapper.java
@@ -1,0 +1,7 @@
+package com.ruleforge.datasource.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.ruleforge.datasource.entity.DatasourceEntityMapping;
+
+public interface DatasourceEntityMappingMapper extends BaseMapper<DatasourceEntityMapping> {
+}