+++ server/lib/ruleforge-datasource/src/main/java/com/ruleforge/datasource/mapper/DatasourceFieldMappingMapper.java
@@ -1,0 +1,7 @@
+package com.ruleforge.datasource.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.ruleforge.datasource.entity.DatasourceFieldMapping;
+
+public interface DatasourceFieldMappingMapper extends BaseMapper<DatasourceFieldMapping> {
+}