+++ server/lib/ruleforge-datasource/src/main/java/com/ruleforge/datasource/mapper/RuleVariableDefMapper.java
@@ -1,0 +1,10 @@
+package com.ruleforge.datasource.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.ruleforge.datasource.entity.RuleVariableDef;
+
+/**
+ * nd_rule_variable_def 的 Mapper 接口
+ */
+public interface RuleVariableDefMapper extends BaseMapper<RuleVariableDef> {
+}